feat: integrate AI layer functionality with new composable
